requirements.txt
Flask-Caching
spacy
Flask-Limiter
